requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages

# apache-airflow-providers-google==10.7.0
astronomer-cosmos[dbt-bigquery]
protobuf